refactor: Restructure Services namespace files
