+++ SyntriceEShop.API/Program.cs
@@ -1,14 +1,12 @@
 ﻿// See https://aka.ms/new-console-template for more information
 
 using SyntriceEShop.API;
-using SyntriceEShop.API.Utilities;
 
 var builder = WebApplication.CreateBuilder(args);
 
 builder.SetupDatabase();
 builder.SetupRepositories();
 builder.SetupServices();
-builder.SetupUtilities();
 
 builder.Services.AddControllers();
 builder.Services.AddEndpointsApiExplorer();
