+++ SyntriceEShop.API/Services/UserServices/IUserService.cs
@@ -1,7 +1,6 @@
-using SyntriceEShop.API.Services.Response;
 using SyntriceEShop.Common.Models.UserModel;
 
-namespace SyntriceEShop.API.Services;
+namespace SyntriceEShop.API.Services.UserServices;
 
 public interface IUserService
 {
