+++ SyntriceEShop.Tests/API/Services/UserServiceTests.cs
@@ -3,8 +3,7 @@ using NSubstitute;
 using Shouldly;
 using SyntriceEShop.API.Repositories;
 using SyntriceEShop.API.Services;
-using SyntriceEShop.API.Services.Response;
-using SyntriceEShop.API.Utilities;
+using SyntriceEShop.API.Services.UserServices;
 using SyntriceEShop.Common.Models.UserModel;
 
 namespace SyntriceEShop.Tests.API.Services;
